Device/Driver/LX/Convert: Restore signed IGC altitudes from LXN
The 7.40 fmt conversion dropped the int16_t casts, so negative
pressure altitudes from Colibri and LX7007 IGC downloads wrapped
to about 65000 m (#3075).

+++ src/Device/Driver/LX/Convert.cpp
@@ -71,8 +71,8 @@ HandlePosition(BufferedOutputStream &os, Context &context,
            /* altitudes can be negative, so cast the uint16_t to
               int16_t to interpret the most significant bit as sign
               bit */
-           FromBE16(position.aalt),
-           FromBE16(position.galt));
+           (int16_t)FromBE16(position.aalt),
+           (int16_t)FromBE16(position.galt));
 
     if (context.b_ext.num == 0)
         os.Write("\r\n");
